# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/posts/{postId}/images": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get all images for a blog post",
                "description": "Retrieve all images associated with a specific blog post.",
                "operationId": "getImages",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of the blog post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of images",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Image"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Images"
                ],
                "summary": "Upload images for a blog post",
                "description": "Upload multiple images for a specific blog post.",
                "operationId": "uploadImages",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of the blog post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Images to upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "images"
                                ],
                                "properties": {
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "description": "Image files to upload",
                                            "type": "file",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Images uploaded successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Images uploaded successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error uploading images",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/images/{id}/primary": {
            "put": {
                "tags": [
                    "Images"
                ],
                "summary": "Set an image as the primary",
                "description": "Set a specific image as the primary (featured) image for a blog post.",
                "operationId": "setPrimary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the image",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Primary image set successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Primary image set successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/images/{id}": {
            "delete": {
                "tags": [
                    "Images"
                ],
                "summary": "Delete an image",
                "description": "Delete a specific image from a blog post.",
                "operationId": "deleteImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the image",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Image deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Image": {
                "description": "Image schema",
                "required": [
                    "id",
                    "url",
                    "is_primary"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the image",
                        "type": "integer",
                        "format": "int64"
                    },
                    "url": {
                        "description": "URL of the image",
                        "type": "string",
                        "format": "uri"
                    },
                    "is_primary": {
                        "description": "Whether the image is the primary one for the post",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}